app/web/requirements.txt
fastapi==0.49.0
uvicorn==0.11.3
SQLAlchemy==1.3.13
typing==3.7.4.1
pydantic==1.4
zeep==3.4.0
suds-py3==1.3.4.0